Switch to raspberry_pi_3

+++ src/common/mem.rs
@@ -1,8 +1,7 @@
-use crate::arch::IntPtr;
 use crate::pointer_iter::PointerIter;
 use core::ptr::write_volatile;
 
-pub unsafe fn zero_region_volatile(range: PointerIter<IntPtr>) {
+pub unsafe fn zero_region_volatile(range: PointerIter<i64>) {
     for ptr in range {
         write_volatile(ptr, 0)
     }
